chore: rename action and pillTrigger component
