--- src/Molecules/Select/Action/index.ts
@@ -1,3 +1,0 @@
-import { Action } from './Action';
-
-export { Action };
