+++ src/Molecules/Select/ActionButton/ActionButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Button, units } from '../../..';
-import { Props } from './Action.types';
+import { Props } from './ActionButton.types';
 
 const StyledButton = styled(Button)`
   padding: ${units(1)}px ${units(3)}px;
@@ -9,6 +9,6 @@ const StyledButton = styled(Button)`
   justify-content: left;
 `;
 
-export function Action(props: Props) {
+export function ActionButton(props: Props) {
   return <StyledButton iconPlacement="left" variant="neutral" fullWidth {...props} />;
 }
