--- src/Molecules/Select/PillTrigger/index.ts
@@ -1,3 +1,0 @@
-import { PillTrigger } from './PillTrigger';
-
-export { PillTrigger };
