+++ src/Molecules/Select/TriggerPill/index.ts
@@ -1,0 +1,3 @@
+import { TriggerPill } from './TriggerPill';
+
+export { TriggerPill };
